Updated to correct variable types

+++ modified_concerts_schema.sql
@@ -1,27 +1,27 @@
 CREATE TABLE artists (
-    artist_id INT PRIMARY KEY,
+    artist_id  INTEGER(10) PRIMARY KEY,
     artist_name VARCHAR(128),
     genre VARCHAR(64),
     debut_year YEAR
 );
 
 CREATE TABLE albums (
-    album_id INT PRIMARY KEY,
+    album_id  INTEGER(10) PRIMARY KEY,
     album_title VARCHAR(256),
     release_date DATE
 );
 
 -- Linking table artists and albums to support multiple artists per album
 CREATE TABLE artist_albums_link (
-    artist_id INT,
-    album_id INT,
+    artist_id  INTEGER(10),
+    album_id  INTEGER(10),
     PRIMARY KEY (artist_id, album_id),
     FOREIGN KEY (artist_id) REFERENCES artists(artist_id),
     FOREIGN KEY (album_id) REFERENCES albums(album_id)
 );
 
 CREATE TABLE songs (
-    song_id INT PRIMARY KEY,
+    song_id  INTEGER(10) PRIMARY KEY,
     song_title VARCHAR(128),
     length REAL,
     release_date DATE
@@ -29,15 +29,15 @@ CREATE TABLE songs (
 
 -- Linking table artists and songs to support multiple artists per song
 CREATE TABLE artists_songs_link (
-    artist_id INT,
-    song_id INT,
+    artist_id  INTEGER(10),
+    song_id  INTEGER(10),
     PRIMARY KEY (artist_id, song_id),
     FOREIGN KEY (artist_id) REFERENCES artists(artist_id),
     FOREIGN KEY (song_id) REFERENCES songs(song_id)
 );
 
 CREATE TABLE concerts (
-    concert_id INT PRIMARY KEY,
+    concert_id  INTEGER(10) PRIMARY KEY,
     concert_title VARCHAR(256),
     location VARCHAR(256),
     date_of_concert DATE
@@ -45,8 +45,8 @@ CREATE TABLE concerts (
 
 -- Linking table artists and concerts to support multiple artists per consert
 CREATE TABLE artist_concerts_link (
-    artist_id INT,
-    concert_id INT,
+    artist_id  INTEGER(10),
+    concert_id  INTEGER(10),
     PRIMARY KEY (artist_id, concert_id),
     FOREIGN KEY (artist_id) REFERENCES artists(artist_id),
     FOREIGN KEY (concert_id) REFERENCES concerts(concert_id)
@@ -54,33 +54,33 @@ CREATE TABLE artist_concerts_link (
 
 -- Linking table concerts and songs to support multiple songs per concert
 CREATE TABLE concerts_songs_link (
-    concert_id INT,
-    song_id INT,
-    order_performance TINYINT,
+    concert_id  INTEGER(10),
+    song_id  INTEGER(10),
+    order_performance TINYINT(4),
     PRIMARY KEY (concert_id, song_id),
     FOREIGN KEY (concert_id) REFERENCES concerts(concert_id),
     FOREIGN KEY (song_id) REFERENCES songs(song_id)
 );
 
 CREATE TABLE fans (
-    fan_id INT PRIMARY KEY,
+    fan_id  INTEGER(10) PRIMARY KEY,
     fan_name VARCHAR(128),
     fan_email VARCHAR(128) UNIQUE,
-    age TINYINT
+    age TINYINT(3)
 );
 
 -- Favorite artists per table
 CREATE TABLE fan_favorites (
-    fan_id INT,
-    artist_id INT,
+    fan_id  INTEGER(10),
+    artist_id  INTEGER(10),
     PRIMARY KEY (fan_id, artist_id),
     FOREIGN KEY (fan_id) REFERENCES fans(fan_id),
     FOREIGN KEY (artist_id) REFERENCES artists(artist_id)
 );
 
 CREATE TABLE concerts_tickets (
-    ticket_id INT PRIMARY KEY,
-    concert_id INT,
+    ticket_id INTEGER(10) PRIMARY KEY,
+    concert_id INTEGER(10),
     purchase_date DATE,
     ticket_price DOUBLE,
     seat_zone VARCHAR(16),
